requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements.in
#
aiofiles==22.1.0
    # via -r requirements.in
anyio==3.6.2
    # via
    #   starlette
    #   watchfiles
certifi==2022.9.24
    # via requests
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via uvicorn
colorama==0.4.5
    # via
    #   click
    #   tqdm
    #   uvicorn
contourpy==1.0.5
    # via matplotlib
cycler==0.11.0
    # via matplotlib
fastapi[standard]==0.85.1
    # via -r requirements.in
fonttools==4.38.0
    # via matplotlib
h11==0.14.0
    # via uvicorn
httptools==0.5.0
    # via uvicorn
idna==3.4
    # via
    #   anyio
    #   requests
kiwisolver==1.4.4
    # via matplotlib
matplotlib==3.6.0
    # via
    #   -r requirements.in
    #   seaborn
numpy==1.22.0 # set manually for dependency conflicts
    # via
    #   contourpy
    #   matplotlib
    #   opencv-python-headless
    #   pandas
    #   seaborn
    #   torchvision
opencv-python-headless==4.6.0.66
    # via -r requirements.in
packaging==21.3
    # via matplotlib
pandas==1.4 # set manually for dependency conflicts
    # via
    #   -r requirements.in
    #   seaborn
pillow==9.2.0
    # via
    #   -r requirements.in
    #   matplotlib
    #   torchvision
psutil==5.9.3
    # set manually
pydantic==1.10.2
    # via fastapi
pyparsing==3.0.9
    # via
    #   matplotlib
    #   packaging
python-dateutil==2.8.2
    # via
    #   matplotlib
    #   pandas
python-dotenv==0.21.0
    # via uvicorn
python-multipart==0.0.5
    # via -r requirements.in
pytz==2022.5
    # via pandas
pyyaml==6.0
    # via uvicorn
requests==2.28.1
    # via
    #   -r requirements.in
    #   torchvision
scipy==1.9.3
    # set manually
seaborn==0.12.1
    # via -r requirements.in
six==1.16.0
    # via
    #   python-dateutil
    #   python-multipart
sniffio==1.3.0
    # via anyio
starlette==0.20.4
    # via fastapi
torch==1.9.1
    # via
    #   -r requirements.in
    #   torchvision
torchvision==0.10.1
    # via -r requirements.in
tqdm==4.64.1
    # via -r requirements.in
typing-extensions==4.4.0
    # via
    #   pydantic
    #   starlette
    #   torch
    #   torchvision
urllib3==1.26.12
    # via requests
uvicorn[standard]==0.19.0
    # via -r requirements.in
watchfiles==0.18.0
    # via uvicorn
websockets==10.3
    # via uvicorn
